+++ HISTORY.rst
@@ -16,6 +16,7 @@ History
 - Disable breadcrumbs when RanaBrowser is disabled
 - Create threedi models via rana process (nens/rana-qgis-plugin#258)
 - Download styling for raster and vector files (nens/rana-qgis-plugin#274)
+- Add cancel button for running simulations (nens/rana-qgis-plugin#235)
 
 
 1.2.4 (2026-01-28)